Aye/requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
automium==0.2.6
automium-web==0.1.1
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
client==0.0.1
clubs==0.1.0
constantly==15.1.0
cryptography==2.9.2
distlib==0.3.0
Django==3.0.5
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
Flask==1.1.2
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.19.0
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
service-identity==18.1.0
signals==0.0.2
six==1.15.0
sqlparse==0.2.4
twilio==6.44.0
urllib3==1.25.9
users==1.0.dev0
views==0.3
virtualenv==20.0.23
Werkzeug==1.0.1
zope.interface==5.1.0
